+++ cypress/integration/screenshot.e2e.js
@@ -1,0 +1,11 @@
+import { viewport } from '../support/cypressHelpers'
+
+describe('Integration test with visual testing', () => {
+  it('loads intial page', () => {
+    cy.server()
+      .viewport(viewport.w, viewport.h)
+      .visit('/')
+      .wait(2000)
+      .percySnapshot()
+  })
+})
