+++ cypress/plugins/index.js
@@ -1,19 +1,7 @@
-/* eslint-disable no-console, @typescript-eslint/explicit-function-return-type */
-// ***********************************************************
-// This example plugins/index.js can be used to load plugins
-//
-// You can change the location of this file or turn off loading
-// the plugins file with the 'pluginsFile' configuration option.
-//
-// You can read more here:
-// https://on.cypress.io/plugins-guide
-// ***********************************************************
+/* eslint-disable no-console, @typescript-eslint/explicit-function-return-type, @typescript-eslint/no-var-requires */
 
-// This function is called when a project is opened or re-opened (e.g. due to
-// the project's config changing)
+const percyHealthCheck = require('@percy/cypress/task')
 
 module.exports = (on, config) => {
-  // `on` is used to hook into various events Cypress emits
-  // `config` is the resolved Cypress config
-  console.log(on, config)
+  on('task', percyHealthCheck)
 }
