+++ cypress/support/cypressHelpers.js
@@ -1,0 +1,12 @@
+const viewport = {
+  w: 550,
+  h: 750,
+}
+
+const section = {
+  home: '[data-test="home"]',
+  about: '[data-test="about"]',
+  contact: '[data-test="contact"]',
+}
+
+export { viewport, section }
