+++ package.json
@@ -28,7 +28,7 @@
     "test:e2e": "NODE_ENV=development yarn build && start-server-and-test serve http-get://localhost:3001/ cypress:open",
     "test:e2e:run": "NODE_ENV=development yarn build && start-server-and-test serve http-get://localhost:3001/ cypress:run",
     "cypress:open": "cypress open",
-    "cypress:run": "cypress run",
+    "cypress:run": "percy exec -- cypress run",
     "--------- HELPERS ----------------------------": "",
     "check:status": "node -e 'require(\"./config/helpers/nodeScripts\").statusCheck()' && node -e 'require(\"./config/helpers/nodeScripts\").branchCheck()'",
     "copy:assets": "node -e 'require(\"./config/helpers/nodeScripts\").moveAssets()'",
@@ -82,6 +82,7 @@
     "@babel/preset-typescript": "^7.3.2",
     "@babel/register": "7.4.4",
     "@babel/runtime": "7.4.5",
+    "@percy/cypress": "^2.0.0",
     "@types/jest": "^24.0.17",
     "@types/react": "^16.9.1",
     "@types/react-dom": "^16.8.5",
